+++ tests/testcases_output/php/ok_interfaces_graphviz/docs.php
@@ -1,0 +1,12 @@
+<?php
+global $pasdoc;
+$pasdoc = array(
+  'ok_interfaces_graphviz' => array('html_filename' => 'ok_interfaces_graphviz.html', 'type' => 'unit'),
+  'TBaseClass' => array('html_filename' => 'ok_interfaces_graphviz.TBaseClass.html', 'type' => 'class'),
+  'TDerivedClass' => array('html_filename' => 'ok_interfaces_graphviz.TDerivedClass.html', 'type' => 'class'),
+  'TDerivedClassWithInterface' => array('html_filename' => 'ok_interfaces_graphviz.TDerivedClassWithInterface.html', 'type' => 'class'),
+  'IMyInterface' => array('html_filename' => 'ok_interfaces_graphviz.IMyInterface.html', 'type' => 'interface'),
+  'IMyInterface2' => array('html_filename' => 'ok_interfaces_graphviz.IMyInterface2.html', 'type' => 'interface'),
+  'TDerivedClassWithMyInterface' => array('html_filename' => 'ok_interfaces_graphviz.TDerivedClassWithMyInterface.html', 'type' => 'class'),
+  'TDerivedClassWithMoreInterfaces' => array('html_filename' => 'ok_interfaces_graphviz.TDerivedClassWithMoreInterfaces.html', 'type' => 'class'),
+);